requirements.txt
alembic==1.6.5
astroid==2.7.3
async-generator==1.10
attrs==21.2.0
bcrypt==3.2.0
blinker==1.6.2
cachelib==0.2.0
certifi==2021.10.8
cffi==1.14.6
click==8.0.1
coverage==6.0.1
cryptography==36.0.1
execnet==1.9.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
greenlet==1.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
outcome==1.1.0
packaging==21.0
platformdirs==2.3.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.7.0
pylint==2.10.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-cov==3.0.0
pytest-forked==1.3.0
pytest-xdist==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.18
tblib==1.7.0
toml==0.10.2
tomli==1.2.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.7.1
urllib3==1.26.8
Werkzeug==2.0.1
wrapt==1.12.1
wsproto==1.1.0
